requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alabaster==1.0.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
auth0-python==4.8.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.13.12
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
crew==0.9.2
crewai==0.102.0
crewai-tools==0.36.0
cryptography==44.0.1
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docutils==0.21.2
durationpy==0.9
embedchain==0.1.126
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
fastavro==1.10.0
fastjsonschema==2.21.1
fica==0.4.1
filelock==3.17.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.80.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipykernel==6.29.5
ipylab==1.0.0
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.1
json5==0.10.0
json_repair==0.36.1
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
jupytext==1.16.7
kiwisolver==1.4.8
kubernetes==32.0.0
lancedb==0.19.0
langchain==0.3.18
langchain-cohere==0.3.5
langchain-community==0.3.17
langchain-core==0.3.35
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.6
langsmith==0.1.147
litellm==1.60.2
lxml==6.0.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mem0ai==0.1.49
mistune==3.1.3
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
nose==1.3.7
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.63.0
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
oscrypto==1.3.0
otter-grader==6.1.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
playwright==1.51.0
plumbum==1.9.0
ply==3.11
portalocker==2.10.1
posthog==3.13.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pylance==0.23.0
pyparsing==3.2.3
pypdf==5.3.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.394
pysbd==0.3.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-on-whales==0.76.1
python3-pika==0.9.14
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.3.0
qdrant-client==1.13.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
shapely==2.0.7
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
svglib==1.5.1
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Tree==0.2.4
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.3.0
uv==0.6.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==1.7.6
xhtml2pdf==0.2.17
yarl==1.18.3
zipp==3.21.0
